added UIButton extensions in a separate file + created classe for special UI elements
