+++ Extensions/Today/TodayViewController.swift
@@ -191,113 +191,3 @@ class TodayViewController: UIViewController, NCWidgetProviding, TodayWidgetAppea
     }
 }
 
-extension UIButton {
-    func setBackgroundColor(_ color: UIColor, forState state: UIControl.State) {
-        let colorView = UIView(frame: CGRect(width: 1, height: 1))
-        colorView.backgroundColor = color
-
-        UIGraphicsBeginImageContext(colorView.bounds.size)
-        if let context = UIGraphicsGetCurrentContext() {
-            colorView.layer.render(in: context)
-        }
-        let colorImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-
-        self.setBackgroundImage(colorImage, for: state)
-    }
-}
-
-class ImageButtonWithLabel: UIView {
-
-    lazy var button = UIButton()
-    lazy var label = UILabel()
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        performLayout()
-    }
-
-    func performLayout() {
-        addSubview(button)
-        addSubview(label)
-        button.imageView?.contentMode = .scaleAspectFit
-
-        button.snp.makeConstraints { make in
-            make.centerX.equalTo(self)
-            make.top.equalTo(self.safeAreaLayoutGuide).offset(5)
-            make.right.greaterThanOrEqualTo(self.safeAreaLayoutGuide).offset(40)
-            make.left.greaterThanOrEqualTo(self.safeAreaLayoutGuide).inset(40)
-            make.height.greaterThanOrEqualTo(60)
-        }
-
-        label.snp.makeConstraints { make in
-            make.top.equalTo(button.snp.bottom).offset(10)
-            make.leading.trailing.bottom.equalTo(self)
-            make.height.equalTo(10)
-        }
-
-        label.numberOfLines = 1
-        label.lineBreakMode = .byWordWrapping
-        label.textAlignment = .center
-    }
-
-    func addTarget(_ target: AnyObject?, action: Selector, forControlEvents events: UIControl.Event) {
-        button.addTarget(target, action: action, for: events)
-    }
-}
-
-class ButtonWithSublabel: UIButton {
-    lazy var subtitleLabel = UILabel()
-    lazy var label = UILabel()
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    convenience init() {
-        self.init(frame: .zero)
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        performLayout()
-    }
-
-    fileprivate func performLayout() {
-        let titleLabel = self.label
-        self.titleLabel?.removeFromSuperview()
-        addSubview(titleLabel)
-
-        let imageView = self.imageView!
-        let subtitleLabel = self.subtitleLabel
-        self.addSubview(subtitleLabel)
-
-        imageView.snp.makeConstraints { make in
-            make.centerY.left.equalTo(10)
-            make.width.equalTo(TodayUX.copyLinkImageWidth)
-        }
-
-        titleLabel.snp.makeConstraints { make in
-            make.left.equalTo(imageView.snp.right).offset(10)
-            make.trailing.top.equalTo(self)
-            make.height.greaterThanOrEqualTo(12)
-        }
-
-        subtitleLabel.lineBreakMode = .byTruncatingTail
-        subtitleLabel.snp.makeConstraints { make in
-            make.bottom.equalTo(self).inset(10)
-            make.top.equalTo(titleLabel.snp.bottom)
-            make.leading.trailing.equalTo(titleLabel)
-            make.height.greaterThanOrEqualTo(10)
-        }
-    }
-
-    override func setTitle(_ text: String?, for state: UIControl.State) {
-        self.label.text = text
-        super.setTitle(text, for: state)
-    }
-}
